added retirement spec and fixed cart transfer on creation of new user

+++ app/models/product.rb
@@ -25,9 +25,9 @@ class Product < ActiveRecord::Base
   end
 
   def retire
-    self.retired = true
-    self.cart_products.destroy_all
-    save
+    if self.update_attribute(:retired, true)
+      self.cart_products.destroy_all
+    end
   end
 
   def active?
